pyproject.toml
[tool.poetry]
name = "goldboost"
version = "0.1.0"
description = ""
authors = ["dhushchyn <dhushchyn@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django-environ = "^0.11.2"
psycopg2-binary = "^2.9.9"
django-unfold = "^0.20.4"
pillow = "^10.2.0"
django-ninja = "^1.1.0"
django-ninja-extra = "^0.20.4"
django-meta = "^2.4.2"
instagram-private-api = "^1.6.0.0"
instaloader = "^4.10.3"
gunicorn = "^21.2.0"
django-cors-headers = "^4.3.1"
pydantic = {extras = ["email"], version = "^2.6.3"}
django-ninja-jwt = "^5.3.0"
celery = "^5.3.6"
redis = "^5.0.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
